Extracts followsController from userController
