+++ semana18/cookenu/src/controllers/userController.ts
@@ -1,17 +1,8 @@
 import { Request, Response } from "express";
-import { UserFollow } from "../types/user";
-import {
-  deleteFollow,
-  dropUser,
-  getFeed,
-  insertFollow,
-  selectUserById,
-} from "../data/userQueries";
+import { dropUser, getFeed, selectUserById } from "../data/userQueries";
 import { tokenValidator } from "../services/tokenService";
 import CustomError from "../errors/customError";
 import { hasHeaderToken } from "../validations/validHeaderToken";
-import { validFollowField } from "../validations/validFieldFollow";
-import { validUnfollowField } from "../validations/validFieldUnfollow";
 import { formatData } from "../util/transformData";
 import { FeedData } from "../types/feed";
 
@@ -43,44 +34,6 @@ export default class UserController {
       const { id, name, email } = result;
 
       res.send({ id, name, email });
-    } catch ({ code, message }) {
-      res.status(code ? code : 400).send({ message });
-    }
-  };
-
-  follow = async (req: Request, res: Response) => {
-    try {
-      const token = hasHeaderToken(req.headers);
-      const { id: follower } = tokenValidator(token);
-      const followed = await validFollowField(req.body);
-
-      const data: UserFollow = {
-        followed_id: followed,
-        follower_id: follower,
-      };
-
-      await insertFollow(data);
-
-      res.send({ message: "Followed" });
-    } catch ({ code, message }) {
-      res.status(code ? code : 400).send({ message });
-    }
-  };
-
-  unfollow = async (req: Request, res: Response) => {
-    try {
-      const token = hasHeaderToken(req.headers);
-      const { id: follower } = tokenValidator(token);
-      const followed = await validUnfollowField(req.body);
-
-      const data: UserFollow = {
-        followed_id: followed,
-        follower_id: follower,
-      };
-
-      await deleteFollow(data);
-
-      res.send({ message: "Unfollowed" });
     } catch ({ code, message }) {
       res.status(code ? code : 400).send({ message });
     }
